Log important changes
During normal operation, firewalld seems to be silent in syslog.
This is an attempt to make some information available for cases
that might matter to the system state.

These might change how log messages of other daemons etc. are to be
interpreted:

* Firewall reached state 'RUNNING'
* Reloaded, reloaded completely
* [For symmetry] Copied runtime to permanent config
* Unloading all firewall-related modules
  [This might / is intended to break existing connections.]
* Finished/skipping CleanupOnExit
  [The state the system/firewall is left at]
* [Warning] Enabling/disabling panic mode

These change how the firewall itself is secured against the system:

* Lockdown enabled/disabled