firewall.core.logger: Fix logic error in _delLogging() by correcting indentation
The continue should happen when a matching entry was removed, and not
just when the list of entries happens to get empty and the key gets
removed also.

* Bonus: Fix typo in exception message.

Here is an example how the behaviour differs.

Before:

  >>> from firewall.core.logger import log
  >>> log.addInfoLogging("*", log.stdout, log.INFO1)
  >>> len(log._logging[log.INFO1])
  2
  >>> log.delInfoLogging("*", log.stdout, log.INFO1)
  Traceback (most recent call last):
    File "<stdin>", line 1, in <module>
    File "[...]/github/firewalld/src/firewall/core/logger.py", line 395, in delInfoLogging
      self._delLogging(domain, target, level, fmt, is_debug=0)
    File "[...]/github/firewalld/src/firewall/core/logger.py", line 579, in _delLogging
      (_level, domain, target.__class__.__name__, fmt))
  ValueError: No mathing logging for level 1, domain *, target _StdoutLog and format None.

After:

  >>> from firewall.core.logger import log
  >>> log.addInfoLogging("*", log.stdout, log.INFO1)
  >>> len(log._logging[log.INFO1])
  2
  >>> log.delInfoLogging("*", log.stdout, log.INFO1)
  >>> len(log._logging[log.INFO1])
  1
  >>> log.delInfoLogging("*", log.stdout, log.INFO1)
  >>> len(log._logging[log.INFO1])
  Traceback (most recent call last):
    File "<stdin>", line 1, in <module>
  KeyError: 1

So, before, it gave an error that no matching logging was found,
although there were 2 of them. After, it first removes one logging,
then the other, and with the last removed the key is gone. This
seems to be more like what was intended.